Failing test for #1005
The files in `bin/` such as `bin/rake` are come from the compiled binaries. For some versions of Ruby they start with a binstub that looks like this:

```
$ cd ~/Downloads/ruby-2.3.8
$ ls bin
erb	gem	irb	rake	rdoc	ri	ruby
$ cat bin/rake | head -n 1
#!/app/vendor/ruby-2.3.8/bin/ruby
```

This does not work at build due to the path `/app` not being where the app is.

This test fails because `bin/rake` has a bad shebang line when `rake assets:precompile` is run it will give an error:

```
       +remote: -----> Detecting rake tasks
       +remote:        Could not detect rake tasks
       +remote:        ensure you can run `$ bundle exec rake -P` against your app
       +remote:        and using the production group of your Gemfile.
       +remote:        bash: /tmp/build_d8662c03cfa0321567ab12e0a1e6f4f3/bin/rake: /app/vendor/ruby-2.4.9/bin/ruby: bad interpreter: No such file or directory
```